+++ checkpoint/orbax/checkpoint/experimental/v1/path/__init__.py
@@ -21,4 +21,4 @@ from orbax.checkpoint.experimental.v1._src.path.types import (
     PathLike,
     PathAwaitingCreation,
 )
-from orbax.checkpoint.experimental.v1._src.path import step
+from orbax.checkpoint.experimental.v1.path import step